+++ tilebox-grpc/_tilebox/grpc/channel.py
@@ -2,6 +2,7 @@ import json
 import re
 from collections.abc import Callable
 from dataclasses import dataclass
+from enum import Enum
 from typing import TypeVar
 
 from grpc import (
@@ -42,11 +43,21 @@ CHANNEL_OPTIONS = [
     ("grpc.service_config", json.dumps(_SERVICE_CONFIG)),
 ]
 
+class ChannelProtocol(Enum):
+    HTTPS = 1
+    HTTP = 2
+    UNIX = 3
+
 
 @dataclass
 class ChannelInfo:
-    url_without_protocol: str
-    use_ssl: bool
+    address: str
+    """GRPC target address. For http(s) connections this is the host[:port] format without a scheme. For unix sockets
+    this is the unix socket path including the `unix://` prefix for absolute paths or `unix:` for relative paths."""
+    port: int
+    """Port number for http(s) connections. For unix sockets this is always 0."""
+    protocol: ChannelProtocol
+    """The protocol to use for the channel."""
 
 
 def open_channel(url: str, auth_token: str | None = None) -> Channel:
@@ -69,14 +80,22 @@ def open_channel(url: str, auth_token: str | None = None) -> Channel:
 
 
 def _open_channel(channel_info: ChannelInfo) -> Channel:
-    if channel_info.use_ssl:
-        return secure_channel(
-            channel_info.url_without_protocol,
-            ssl_channel_credentials(),
-            CHANNEL_OPTIONS,
-            compression=Compression.Gzip,
-        )
-    return insecure_channel(channel_info.url_without_protocol, CHANNEL_OPTIONS, compression=Compression.NoCompression)
+    match channel_info.protocol:
+        case ChannelProtocol.HTTPS:
+            return secure_channel(
+                f"{channel_info.address}:{channel_info.port}",
+                ssl_channel_credentials(),
+                CHANNEL_OPTIONS,
+                compression=Compression.Gzip,
+            )
+        case ChannelProtocol.HTTP:
+            return insecure_channel(
+                f"{channel_info.address}:{channel_info.port}", CHANNEL_OPTIONS, compression=Compression.NoCompression
+            )
+        case ChannelProtocol.UNIX:
+            return insecure_channel(channel_info.address, CHANNEL_OPTIONS, compression=Compression.NoCompression)
+        case _:
+            raise ValueError(f"Unsupported channel protocol: {channel_info.protocol}")
 
 
 _URL_SCHEME = re.compile(r"^(https?://)?([^: ]+)(:\d+)?/?$")
@@ -98,27 +117,28 @@ def parse_channel_info(url: str) -> ChannelInfo:
         A ChannelInfo object that can be used to create a gRPC channel.
     """
     # See https://github.com/grpc/grpc/blob/master/doc/naming.md
-    if url.startswith("unix:"):
-        return ChannelInfo(url, False)
+    if url.startswith("unix:"):  ## unix:///absolute/path or unix://path
+        return ChannelInfo(url, 0, ChannelProtocol.UNIX)
 
     # `urllib.parse.urlparse` behaves a bit weird with URLs that don't have a scheme but a port number, so regex it is
     if (match := _URL_SCHEME.match(url)) is None:
         raise ValueError(f"Invalid URL: {url}")
     scheme, netloc, port = match.groups()
     netloc = netloc.rstrip("/")
-    use_ssl = True
+    protocol = ChannelProtocol.HTTPS
 
     if scheme == "http://":  # explicitly set http -> require a port
         if port is None:
             raise ValueError("Explicit port required for insecure HTTP channel")
-        use_ssl = False
+        protocol = ChannelProtocol.HTTP
 
-    if scheme is None and port is not None:  # no scheme, but a port that looks like a dev port -> insecure
-        use_ssl = port == ":443"
+    # no scheme, but a port that looks like a dev port -> insecure
+    if scheme is None and port is not None and port != ":443":
+        protocol = ChannelProtocol.HTTP
 
-    if use_ssl:
-        return ChannelInfo(netloc + (port or ":443"), True)
-    return ChannelInfo(netloc + port, False)
+    port_number = 443 if port is None else int(port.removeprefix(":"))
+
+    return ChannelInfo(netloc, port_number, protocol)
 
 
 RequestType = TypeVar("RequestType")