+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- `tilebox-grpc`: More robust parsing of GRPC channel URLs.
+
 ## [0.37.0] - 2025-06-06
 
 ### Changed